version update & fix compile error only

+++ cap-db/src/main/java/com/iisigroup/cap/jdbc/CapNamedJdbcTemplate.java
@@ -268,7 +268,7 @@ public class CapNamedJdbcTemplate extends NamedParameterJdbcTemplate {
         }
         long cur = System.currentTimeMillis();
         try {
-            return super.queryForObject(sql.toString(), (Map) args, Integer.class);
+            return (int) super.queryForObject(sql.toString(), (Map) args, Integer.class);
         } catch (Exception e) {
             throw new CapDBException(e, causeClass);
         } finally {
